# PERF: groupby with string dtype