@simonjayhawkins 2021-09-18
does this also fix the case mentioned in https://github.com/pandas-dev/pandas/issues/41596#issuecomment-854628297? if so can you post the timings for that too for df_str.groupby('a')[cols[1:]].last() on master and with this PR