@debnathshoham 2021-09-18
Hi @simonjayhawkins - no this PR doesn't address the performance slowdown in the `object` dtype.
It only ensures that `string` takes the cython-path (similar to the `object` dtype).

Also note, that I have changed the size of the `df` in the top comment, wrt the example in the bug report.