@simonjayhawkins 2021-09-18
That's okay, but we can't close that issue. will need to open a new one specifically for that case, (regression between 1.2.5 and 1.3) or change the OP to read "xref" instead of "closes"

out of curiosity, did you identify which commit caused the slowdown between 1.1.5 and 1.2?