@debnathshoham 2021-09-18
Updated the top comment. 
Honestly, I didn't try to find the commit. This fix was suggested in a comment in the bug report.