@simonjayhawkins 2021-09-18
So why do we need to special case an specific EA dtype? (Maybe this was done before). In general EA handling should be more generic.