@debnathshoham 2021-09-18
I think the current implementation of `_ea_wrap_cython_operation` works that way, no?

I can see a TODO for a more generic approach.