@simonjayhawkins 2021-09-18
yes, it appears that there is intent to make `_ea_wrap_cython_operation` more generic. I'm just wondering why it worked in 1.1.5