@simonjayhawkins 2021-09-22
> That's okay, but we can't close that issue. will need to open a new one specifically for that case, (regression between 1.2.5 and 1.3) or change the OP to read "xref" instead of "closes"

changed back to "closes" xref https://github.com/pandas-dev/pandas/issues/41596#issuecomment-924080060