@jreback 2021-09-22
> groupby.String.time_str_func: setup: wrong number of arguments (for <bound method String.setup of <benchmarks.groupby.String object at 0x7f6c8c634a30>> in groupby.py:626): expected 2, has 1

in the asvs