@jreback 2021-09-22
also merge master as the deprecation warnings on asv's are silenced now